// src/types/user.type.ts
import { AutoGeneratedField, SearchParams, SimpleResponse, PaginatedResponse } from './common';
import { Status } from '../constants/status';

export type UserFormData = {
  name: string;
  email: string;
  username: string;
  language?: string;
  password?: string;
  roleId?: number;
  status?: Status;
};

export type User = AutoGeneratedField & UserFormData;

// User response types sử dụng User entity
export type UserSimpleResponse = SimpleResponse<User>;
export type UserPaginatedResponse = PaginatedResponse<User>;

export type UserParams = SearchParams<UserFormData> & {};